+++ src/PluginDefinition.h
@@ -36,7 +36,7 @@ const TCHAR NPP_PLUGIN_NAME[] = TEXT("Emoji Description");
 //
 // Here define the number of your plugin commands
 //
-const int nbFunc = 8;
+const int nbFunc = 9;
 
 //
 // Which information fields are shown in the status bar.